requirements.txt
asgiref==3.8.1
Django==5.0.7
gunicorn==22.0.0
packaging==24.1
pillow==10.4.0
pytesseract==0.3.10
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
